moved decompose clifford gates
