+++ tangelo/linq/helpers/circuits/tests/test_clifford_circuits.py
@@ -3,39 +3,16 @@ import unittest
 from math import pi as pi
 from tangelo.linq import Gate, Circuit
 
-from tangelo.linq.helpers.circuits import get_clifford_circuit, decompose_gate_to_cliffords
+from tangelo.linq.helpers.circuits.clifford_circuits import decompose_gate_to_cliffords
 
-gates_1 = [Gate("H", 2), Gate("CNOT", 1, control=0),
-           Gate("RZ", 0, parameter = -0.1 , is_variational=True),
-           Gate("CNOT", 2, control = 1),
-           Gate("RX", 0, parameter = -2.9 * pi, is_variational=True),
-           Gate("RY", 0, parameter = 9.2*pi , is_variational=True),
-           Gate("PHASE", 2, parameter = 0.1 , is_variational=True)]
+gates_decomp = Gate("RX", target=0, parameter=pi / 2)#, Gate("RY", target=1, parameter=-pi / 2)]
 
-circuit_1 = Circuit()
-for gate in gates_1:
-    circuit_1.add_gate(gate)
-
-gates_ref = [Gate("H", 2), Gate("CNOT", 1, control=0),
-            Gate("RZ", 0, parameter = 0 , is_variational=True),
-            Gate("CNOT", 2, control = 1),
-            Gate("RX", 0, parameter = -pi, is_variational=True),
-            Gate("RY", 0, parameter = pi , is_variational=True),
-            Gate("PHASE", 2, parameter = 0 , is_variational=True)]
-
-circuit_ref = Circuit()
-for gate in gates_ref:
-    circuit_ref.add_gate(gate)
+gates_ref_decomp = [Gate("SDAG", 0), Gate("H", 0), Gate("SDAG", 0)]
+                    #Gate("Z", 1), Gate("H", 1)]
 
 class CliffordCircuitTest(unittest.TestCase):
 
-    def test_get_clifford_circuit(self):
-        cliff_circuit = get_clifford_circuit(circuit_1)
-        self.assertEqual(cliff_circuit, circuit_ref)
+    def test_decompose_gate_to_cliffords(self):
+        self.assertEqual(gates_ref_decomp, decompose_gate_to_cliffords(gates_decomp))
+    #def test_non_clifford_gates(self):
 
-
-   # def test_decompose_clifford_gate(self):
-        """Test initializing and uncomputing circuits with cirq lsq_first order"""
-        #im = get_backend("cirq")
-
-      #  np.testing.assert_array_almost_equal(clifford_state, ref_state)
